requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-braces==1.14.0
django-oauth-toolkit==1.5.0
django-rest-framework-social-oauth2==1.1.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
idna==2.10
jwcrypto==0.8
numpy==1.20.2
oauthlib==3.1.0
opencv-python==4.5.1.48
pandas==1.2.4
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
pyzbar==0.1.8
qrcode==6.1
qrtools==0.0.2
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
urllib3==1.26.4
